Similaritor test
Former-commit-id: a0f7b1739a202c08f6b441db8ace664422feba77

+++ record_classification/src/test/java/uk/ac/standrews/cs/digitising_scotland/record_classification/classifiers/lookup/StringLengthSimilarityMetric.java
@@ -11,7 +11,7 @@ class StringLengthSimilarityMetric implements SimilarityMetric<String> {
         return 1/(1+ getLengthDiff(o1,o2));
     }
 
-    protected int getLengthDiff(String o1, String o2) {
+    protected double getLengthDiff(String o1, String o2) {
         return Math.abs(o1.length()-o2.length());
     }
 }
